requirements.txt
google-generativeai
Pillow
python-dotenv
openai
streamlit
requests
fastapi
python-docx
pymupdf
